opp and meeting activation test in pytest
